code/requirements.txt
# Core dependencies
flask>=2.3.0
numpy>=1.24.0
scipy>=1.10.0

# Audio processing
sounddevice>=0.4.6
pydub>=0.25.1
librosa>=0.10.1
SpeechRecognition>=3.14.0

# Machine learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Google Cloud APIs
google-cloud-speech>=2.20.0
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.3.0
google-auth>=2.20.0

# Hardware (Raspberry Pi)
RPi.GPIO>=0.7.1
smbus2>=0.4.0
smbus>=1.1

# Utilities
keyboard>=0.13.5
pyaudio>=0.2.11